requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Video processing
opencv-python>=4.8.0
numpy>=1.26.0
Pillow>=10.1.0

# HTTP client
httpx>=0.25.0
aiofiles>=23.2.0

# LLM providers
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# WhatsApp automation
playwright>=1.40.0

# Utils
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
